node/03-api-solid/src/services/search-gyms.ts
import { Gym } from '@prisma/client'
import { GymsRepository } from '@/repositories/gyms-repository'

interface SearchGymServiceRequest {
  query: string
  page: number
}
interface SearchGymServiceResponse {
  gyms: Gym[]
}
export class SearchGymService {
  constructor(private gymRepository: GymsRepository) {}
  async execute({
    query,
    page,
  }: SearchGymServiceRequest): Promise<SearchGymServiceResponse> {
    const gyms = await this.gymRepository.searchMany(query, page)
    return { gyms }
  }
}
